requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
